Remove a workaround for a bug in `gcloud`
I got a job at Google and fixed the bug, so now I can remove this
workaround.

The bug was that `gcloud functions deploy ...` would fail if all of the
following were true:

1. The user is deploying files from a local filesystem (as opposed to a
Google Cloud Storage bucket).
2. At least one of the files being deployed is not writable by the user
(e.g., because it is owned by root).
3. One of the non-writable files has an `mtime` older than 1980 (e.g.,
midnight on 1970-01-01).

That seems like an unlikely combination of factors, but it's fairly
common when using [Toast](https://github.com/stepchowfun/toast), which
copies files into a Docker container as root without preserving `mtime`s.